backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography
python-dotenv==1.0.0
transformers==4.36.0
torch==2.2.0
pillow==10.1.0
numpy==1.26.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
